+++ subsystems/posix/userland/smoke.c
@@ -101,6 +101,7 @@ int main(void) {
     const char* timertest_argv[] = {"/disk/bin/timertest", 0};
     const char* sectiontest_argv[] = {"/disk/bin/sectiontest", 0};
     const char* semaphoretest_argv[] = {"/disk/bin/semaphoretest", 0};
+    const char* cliptest_argv[] = {"/disk/bin/cliptest", 0};
     const char* mmaptest_argv[] = {"/disk/bin/mmaptest", 0};
     const char* heaptest_argv[] = {"/disk/bin/heaptest", 0};
     const char* sh_argv[] = {"/bin/sh", "-c", "echo busybox-shell > /disk/smoke/sh.txt", 0};
@@ -123,6 +124,7 @@ int main(void) {
         !file_exists("/disk/bin/timertest") ||
         !file_exists("/disk/bin/sectiontest") ||
         !file_exists("/disk/bin/semaphoretest") ||
+        !file_exists("/disk/bin/cliptest") ||
         !file_exists("/disk/bin/mmaptest") ||
         !file_exists("/disk/bin/heaptest") ||
         !file_exists("/disk/bin/busybox") ||
@@ -148,6 +150,7 @@ int main(void) {
         !run_and_expect("/disk/bin/timertest", timertest_argv, 1, 0) ||
         !run_and_expect("/disk/bin/sectiontest", sectiontest_argv, 1, 0) ||
         !run_and_expect("/disk/bin/semaphoretest", semaphoretest_argv, 1, 0) ||
+        !run_and_expect("/disk/bin/cliptest", cliptest_argv, 1, 0) ||
         !run_and_expect("/disk/bin/mmaptest", mmaptest_argv, 1, 0) ||
         !run_and_expect("/disk/bin/heaptest", heaptest_argv, 1, 0) ||
         !prepare_smoke_directory() ||